[main] Move process_stacktrace outside of platformio
