+++ tests/unit_tests/conftest.py
@@ -79,7 +79,7 @@ def mock_run_platformio_cli_run() -> Generator[Mock, None, None]:
 @pytest.fixture
 def mock_decode_pc() -> Generator[Mock, None, None]:
     """Mock _decode_pc for platformio_api."""
-    with patch("esphome.platformio_api._decode_pc") as mock:
+    with patch("esphome.components.esp32._decode_pc") as mock:
         yield mock
